reformat code

+++ src/backoffice/main/com/bank/backoffice/accounts/application/AccountResponse.java
@@ -6,6 +6,6 @@ import com.bank.shared.domain.bus.query.Response;
 
 public record AccountResponse(String id, String number, Double amount) implements Response {
     public static AccountResponse fromAggregate(Account account) {
-        return new AccountResponse(account.id().value(), account.number().value(), account.amount().value());
+        return new AccountResponse(account.id().value(), account.number().value(), account.amount().value().doubleValue());
     }
 }
